Currently, the two-phase flow, homogeneous, without capillary pressure and buoyancy effects works perfectly.
Also, the movie for the displacement of the non-wetting phase has been
written.

In order to solve the linear system, is necessary to use the vertical
values of the total upstream mobilities without the two ghost columns.
Is preferable to change this in order to achieve uniformity in the code.